requirements.txt
#pip install -r requirements.txt
playwright==1.39.0
requests
beautifulsoup4
python-dotenv
httpx
parsel
#playwright-stealth
#setuptools